+++ app/Controllers/UsersController.php
@@ -69,7 +69,6 @@ class UsersController {
     {
       $em = em();
       $user = $em->find(User::class, $id); 
-      $storage = new StorageController;
 
       $user->setName(request()->get('username'))
         ->setEmail(request()->get('email'));
@@ -84,11 +83,14 @@ class UsersController {
 
       request()->get('role') ? $user->setRole(request()->get('role')) : '';
 
-      if ($user->getImage() != null) {
-        $storage->deleteImage($user->getImage());
+      if ($_FILES["image"]["name"] != '') {
+        $storage = new StorageController;
+        if ($user->getImage() != null) {
+          $storage->deleteImage($user->getImage());
+        }
+        
+        $user->setImage($storage->getImage('user', $id));
       }
-      
-      $user->setImage($storage->getImage('user', $id));
 
       $em->merge($user);
       $em->flush();
